# shared/schemas/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Узел 12 — Backend",
    "description": "Этапы 4–5: фоновые планы, принятие, история и CSV.",
    "version": "0.5.0"
  },
  "paths": {
    "/api/login": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Login",
        "description": "Sets an HttpOnly session cookie and returns the CSRF token for later commands.",
        "operationId": "login_api_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionInfo"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "429": {
            "description": "Too Many Requests",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/logout": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Logout",
        "description": "Revokes the session in PostgreSQL and clears the cookie. Idempotent without a session.",
        "operationId": "logout_api_logout_post",
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/me": {
      "get": {
        "tags": [
          "auth"
        ],
        "summary": "Me",
        "operationId": "me_api_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionInfo"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/replans": {
      "post": {
        "summary": "Request Replan",
        "operationId": "request_replan_api_replans_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReplanCommand"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReplanAccepted"
                }
              }
            }
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Service Unavailable"
          }
        }
      },
      "get": {
        "summary": "List Jobs",
        "operationId": "list_jobs_api_replans_get",
        "parameters": [
          {
            "name": "run_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 128,
              "title": "Run Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 50,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ReplanJob"
                  },
                  "title": "Response List Jobs Api Replans Get"
                }
              }
            }
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Service Unavailable"
          }
        }
      }
    },
    "/api/replans/{job_id}": {
      "get": {
        "summary": "Get Job",
        "operationId": "get_job_api_replans__job_id__get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReplanJob"
                }
              }
            }
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Service Unavailable"
          }
        }
      }
    },
    "/api/plans/{plan_id}": {
      "get": {
        "summary": "Get Plan",
        "operationId": "get_plan_api_plans__plan_id__get",
        "parameters": [
          {
            "name": "plan_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Plan Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanResponse"
                }
              }
            }
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Service Unavailable"
          }
        }
      }
    },
    "/api/plans/{plan_id}/apply": {
      "post": {
        "summary": "Accept Plan",
        "operationId": "accept_plan_api_plans__plan_id__apply_post",
        "parameters": [
          {
            "name": "plan_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Plan Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApplyPlanCommand"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandResult"
                }
              }
            }
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Service Unavailable"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ApiError": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "details": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Details"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ApiError"
      },
      "ApplyPlanCommand": {
        "properties": {
          "command_id": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9_.:-]+$",
            "title": "Command Id"
          },
          "run_id": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9_.:-]+$",
            "title": "Run Id"
          },
          "expected_state_version": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Expected State Version"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "command_id",
          "run_id",
          "expected_state_version"
        ],
        "title": "ApplyPlanCommand"
      },
      "Assignment": {
        "properties": {
          "operation_id": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9_.:-]+$",
            "title": "Operation Id"
          },
          "start_s": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Start S"
          },
          "end_s": {
            "type": "integer",
            "minimum": 0.0,
            "title": "End S"
          },
          "track_id": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 128,
                "minLength": 1,
                "pattern": "^[A-Za-z0-9_.:-]+$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Track Id"
          },
          "route_id": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 128,
                "minLength": 1,
                "pattern": "^[A-Za-z0-9_.:-]+$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Route Id"
          },
          "resource_ids": {
            "items": {
              "type": "string",
              "maxLength": 128,
              "minLength": 1,
              "pattern": "^[A-Za-z0-9_.:-]+$"
            },
            "type": "array",
            "title": "Resource Ids"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "operation_id",
          "start_s",
          "end_s"
        ],
        "title": "Assignment"
      },
...
      "CommandResult": {
        "properties": {
          "command_id": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9_.:-]+$",
            "title": "Command Id"
          },
          "run_id": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9_.:-]+$",
            "title": "Run Id"
          },
          "state_version": {
            "type": "integer",
            "minimum": 0.0,
            "title": "State Version"
          },
          "replan_required": {
            "type": "boolean",
            "title": "Replan Required",
            "default": false
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "command_id",
          "run_id",
          "state_version"
        ],
        "title": "CommandResult"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LoginRequest": {
        "properties": {
          "username": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Username",
            "examples": [
              "dispatcher"
            ]
          },
          "password": {
            "type": "string",
            "maxLength": 1024,
            "minLength": 1,
            "format": "password",
            "title": "Password",
            "writeOnly": true
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "username",
          "password"
        ],
        "title": "LoginRequest"
      },
...
      "Plan": {
        "properties": {
          "id": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9_.:-]+$",
            "title": "Id"
          },
          "run_id": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9_.:-]+$",
            "title": "Run Id"
          },
          "based_on_version": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Based On Version"
          },
          "strategy": {
            "type": "string",
            "enum": [
              "passenger_first",
              "earliest_departure"
            ],
            "title": "Strategy"
          },
          "status": {
            "type": "string",
            "enum": [
              "feasible",
              "partial",
              "infeasible",
              "timeout"
            ],
            "title": "Status"
          },
          "assignments": {
            "items": {
              "$ref": "#/components/schemas/Assignment"
            },
            "type": "array",
            "title": "Assignments"
          },
          "unassigned": {
            "items": {
              "$ref": "#/components/schemas/Unassigned"
            },
            "type": "array",
            "title": "Unassigned"
          },
          "metrics": {
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ]
            },
            "type": "object",
            "title": "Metrics"
          },
          "explanations": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Explanations"
          },
          "timed_out": {
            "type": "boolean",
            "title": "Timed Out",
            "default": false
          },
          "based_on_time_s": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Based On Time S",
            "default": 0
          },
          "calculation_time_ms": {
            "type": "number",
            "minimum": 0.0,
            "title": "Calculation Time Ms",
            "default": 0
          },
          "violations": {
            "items": {
              "$ref": "#/components/schemas/PlanViolation"
            },
            "type": "array",
            "title": "Violations"
          },
          "changes": {
            "items": {
              "$ref": "#/components/schemas/AssignmentChange"
            },
            "type": "array",
            "title": "Changes"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "run_id",
          "based_on_version",
          "strategy",
          "status",
          "assignments",
          "unassigned"
        ],
        "title": "Plan"
      },
      "PlanResponse": {
        "properties": {
          "plan": {
            "$ref": "#/components/schemas/Plan"
          },
          "stale": {
            "type": "boolean",
            "title": "Stale"
          },
          "applicable": {
            "type": "boolean",
            "title": "Applicable"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "plan",
          "stale",
          "applicable"
        ],
        "title": "PlanResponse"
      },
...
      "ReplanAccepted": {
        "properties": {
          "command_id": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9_.:-]+$",
            "title": "Command Id"
          },
          "run_id": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9_.:-]+$",
            "title": "Run Id"
          },
          "job_id": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9_.:-]+$",
            "title": "Job Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "command_id",
          "run_id",
          "job_id"
        ],
        "title": "ReplanAccepted"
      },
      "ReplanCommand": {
        "properties": {
          "command_id": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9_.:-]+$",
            "title": "Command Id"
          },
          "run_id": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9_.:-]+$",
            "title": "Run Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "run_id"
        ],
        "title": "ReplanCommand"
      },
      "ReplanJob": {
        "properties": {
          "job_id": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9_.:-]+$",
            "title": "Job Id"
          },
          "run_id": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9_.:-]+$",
            "title": "Run Id"
          },
          "based_on_version": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Based On Version"
          },
          "based_on_time_s": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Based On Time S"
          },
          "status": {
            "type": "string",
            "enum": [
              "queued",
              "running",
              "completed",
              "failed",
              "superseded"
            ],
            "title": "Status"
          },
          "plan_ids": {
            "items": {
              "type": "string",
              "maxLength": 128,
              "minLength": 1,
              "pattern": "^[A-Za-z0-9_.:-]+$"
            },
            "type": "array",
            "title": "Plan Ids"
          },
          "stale": {
            "type": "boolean",
            "title": "Stale",
            "default": false
          },
          "identical": {
            "type": "boolean",
            "title": "Identical",
            "default": false
          },
          "error": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "string"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "job_id",
          "run_id",
          "based_on_version",
          "based_on_time_s",
          "status"
        ],
        "title": "ReplanJob"
      },
...
      "SessionInfo": {
        "properties": {
          "username": {
            "type": "string",
            "title": "Username"
          },
          "role": {
            "type": "string",
            "enum": [
              "viewer",
              "dispatcher",
              "admin"
            ],
            "title": "Role"
          },
          "permissions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Permissions"
          },
          "expires_at": {
            "type": "string",
            "format": "date-time",
            "title": "Expires At"
          },
          "csrf_token": {
            "type": "string",
            "title": "Csrf Token",
            "description": "Отправляйте в заголовке X-CSRF-Token в каждом POST/PATCH/DELETE. Храните только в памяти страницы; после перезагрузки — GET /api/me."
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "username",
          "role",
          "permissions",
          "expires_at",
          "csrf_token"
        ],
        "title": "SessionInfo"
      },
...
      "Unassigned": {
        "properties": {
          "operation_id": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9_.:-]+$",
            "title": "Operation Id"
          },
          "code": {
            "type": "string",
            "enum": [
              "TRACK_CLOSED",
              "TRACK_OCCUPIED",
              "ROUTE_BUSY",
              "RESOURCE_UNAVAILABLE",
              "PREDECESSOR_INCOMPLETE",
              "NO_FEASIBLE_SLOT",
              "STALE_PLAN"
            ],
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "operation_id",
          "code",
          "message"
        ],
        "title": "Unassigned"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}